+++ lib/neo4j.rb
@@ -71,6 +71,7 @@ require 'neo4j/active_node/query/query_proxy_methods_of_mass_updating'
 require 'neo4j/active_node/query/query_proxy_enumerable'
 require 'neo4j/active_node/query/query_proxy_find_in_batches'
 require 'neo4j/active_node/query/query_proxy_eager_loading'
+require 'neo4j/active_node/query/query_proxy_eager_loading/association_tree'
 require 'neo4j/active_node/query/query_proxy_link'
 require 'neo4j/active_node/labels/index'
 require 'neo4j/active_node/labels/reloading'
